// src/app/layout/backoffice/components/transaction/CargaMasiva/CargaMasiva-List/DTOs/CargaMasiva.dto.ts
export class AddCargaMasiva {
  canalVenta: number;
  puntoVenta: number;
  file: File;
  tdr: string;
  contratante: string;
}
export class SearchCargaMasivaDto {
  canal: string;
  puntoVenta: string;
  idUsuario: number;
  fechaInicio: string;
  fechaFin: string;
  idProceso: number;
  idEstado: number;
}
export class SerchCargaMasivaHisDto {
  idPoliza: number;
}
export class DescargarFactura {

}
export class DescargarExcel {

}